% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{as_nanoarrow_array}
\alias{as_nanoarrow_array}
\alias{from_nanoarrow_array}
\title{Convert an object to a nanoarrow array}
\usage{
as_nanoarrow_array(x, ..., schema = NULL)

from_nanoarrow_array(array, to = NULL, ...)
}
\arguments{
\item{x}{An object to convert to a array}

\item{...}{Passed to S3 methods}

\item{schema}{An optional schema used to enforce conversion to a particular
type. Defaults to \code{\link[=infer_nanoarrow_schema]{infer_nanoarrow_schema()}}.}

\item{array}{An object of class 'nanoarrow_array'}

\item{to}{A target prototype object describing the type to which \code{array}
should be converted.}
}
\value{
An object of class 'nanoarrow_array'
}
\description{
Convert an object to a nanoarrow array
}
