% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as_nanoarrow_buffer}
\alias{as_nanoarrow_buffer}
\title{Convert an object to a nanoarrow buffer}
\usage{
as_nanoarrow_buffer(x, ...)
}
\arguments{
\item{x}{An object to convert to a buffer}

\item{...}{Passed to S3 methods}
}
\value{
An object of class 'nanoarrow_buffer'
}
\description{
Convert an object to a nanoarrow buffer
}
\examples{
array <- as_nanoarrow_array(c(NA, 1:4))
array$buffers
as.raw(array$buffers[[1]])
as.raw(array$buffers[[2]])
convert_buffer(array$buffers[[1]])
convert_buffer(array$buffers[[2]])

}
