% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{as_nanoarrow_schema}
\alias{as_nanoarrow_schema}
\alias{infer_nanoarrow_schema}
\title{Convert an object to a nanoarrow schema}
\usage{
as_nanoarrow_schema(x, ...)

infer_nanoarrow_schema(x, ...)
}
\arguments{
\item{x}{An object to convert to a schema}

\item{...}{Passed to S3 methods}
}
\value{
An object of class 'nanoarrow_schema'
}
\description{
Convert an object to a nanoarrow schema
}
