% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{as_nanoarrow_vctr}
\alias{as_nanoarrow_vctr}
\alias{nanoarrow_vctr}
\title{Experimental Arrow encoded arrays as R vectors}
\usage{
as_nanoarrow_vctr(x, ..., schema = NULL)

nanoarrow_vctr(schema = NULL)
}
\arguments{
\item{x}{An object that works with \code{\link[=as_nanoarrow_array_stream]{as_nanoarrow_array_stream()}}. Most
spatial objects in R already work with this method.}

\item{...}{Passed to \code{\link[=as_nanoarrow_array_stream]{as_nanoarrow_array_stream()}}}

\item{schema}{An optional \code{schema}}
}
\value{
A vctr of class 'nanoarrow_vctr'
}
\description{
Experimental Arrow encoded arrays as R vectors
}
\examples{
array <- as_nanoarrow_array(1:5)
as_nanoarrow_vctr(array)

}
