% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-array-stream.R
\name{convert_array_stream}
\alias{convert_array_stream}
\title{Convert an Array Stream into an R vector}
\usage{
convert_array_stream(array_stream, to = NULL)
}
\arguments{
\item{array_stream}{A \link[=as_nanoarrow_array_stream]{nanoarrow_array_stream}.}

\item{to}{A target prototype object describing the type to which \code{array}
should be converted, or \code{NULL} to use the default conversion.}
}
\value{
An R vector of type \code{to}.
}
\description{
Convert an Array Stream into an R vector
}
\examples{
stream <- as_nanoarrow_array_stream(data.frame(x = 1:5))
str(convert_array_stream(stream))
str(convert_array_stream(stream, to = data.frame(x = double())))

}
