% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{nanoarrow_extension_array}
\alias{nanoarrow_extension_array}
\title{Create Arrow extension arrays}
\usage{
nanoarrow_extension_array(
  storage_array,
  extension_name,
  extension_metadata = NULL
)
}
\arguments{
\item{storage_array}{A \link[=as_nanoarrow_array]{nanoarrow_array}.}

\item{extension_name}{For \code{\link[=na_extension]{na_extension()}}, the extension name. This is
typically namespaced separated by dots (e.g., nanoarrow.r.vctrs).}

\item{extension_metadata}{A string or raw vector defining extension metadata.
Most Arrow extension types define extension metadata as a JSON object.}
}
\value{
A \link[=as_nanoarrow_array]{nanoarrow_array} with attached extension
schema.
}
\description{
Create Arrow extension arrays
}
\examples{
nanoarrow_extension_array(1:10, "some_ext", '{"key": "value"}')

}
