% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointers.R
\name{nanoarrow_pointer_is_valid}
\alias{nanoarrow_pointer_is_valid}
\alias{nanoarrow_pointer_addr_chr}
\alias{nanoarrow_pointer_release}
\alias{nanoarrow_pointer_move}
\alias{nanoarrow_allocate_schema}
\alias{nanoarrow_allocate_array}
\alias{nanoarrow_allocate_array_stream}
\title{Danger zone: low-level pointer operations}
\usage{
nanoarrow_pointer_is_valid(ptr)

nanoarrow_pointer_addr_chr(ptr)

nanoarrow_pointer_release(ptr)

nanoarrow_pointer_move(ptr_src, ptr_dst)

nanoarrow_allocate_schema()

nanoarrow_allocate_array()

nanoarrow_allocate_array_stream()
}
\arguments{
\item{ptr, ptr_src, ptr_dst}{An external pointer to a \verb{struct ArrowSchema},
\verb{struct ArrowArray}, or \verb{struct ArrowArrayStream}.}
}
\description{
Danger zone: low-level pointer operations
}
