% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoarrow-package.R
\name{nanoarrow_version}
\alias{nanoarrow_version}
\alias{nanoarrow_with_zstd}
\alias{nanoarrow_with_lz4}
\title{Underlying 'nanoarrow' C library build}
\usage{
nanoarrow_version(runtime = TRUE)

nanoarrow_with_zstd()

nanoarrow_with_lz4()
}
\arguments{
\item{runtime}{Compare TRUE and FALSE values to detect a
possible ABI mismatch.}
}
\value{
A string identifying the version of nanoarrow this package
was compiled against.
}
\description{
Underlying 'nanoarrow' C library build
}
\examples{
nanoarrow_version()
nanoarrow_with_zstd()
nanoarrow_with_lz4()
}
