% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{read_nanoarrow_array_stream}
\alias{read_nanoarrow_array_stream}
\alias{example_ipc_stream}
\title{Read serialized streams of Arrow data}
\usage{
read_nanoarrow_array_stream(x, ...)

example_ipc_stream()
}
\arguments{
\item{x}{A \code{raw()} vector or connection from which to read binary data}

\item{...}{Passed to S3 methods}
}
\value{
A \link[=as_nanoarrow_array_stream]{nanoarrow_array_stream}
}
\description{
Read serialized streams of Arrow data
}
\examples{
read_nanoarrow_array_stream(example_ipc_stream())

}
