% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{as_nanoarrow_array}
\alias{as_nanoarrow_array}
\title{Convert an object to a nanoarrow array}
\usage{
as_nanoarrow_array(x, ...)
}
\arguments{
\item{x}{An object to convert to a array}

\item{...}{Passed to S3 methods}
}
\value{
An object of class 'nanoarrow_array'
}
\description{
Convert an object to a nanoarrow array
}
