% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-array.R
\name{convert_array}
\alias{convert_array}
\title{Materialize an Array into an R vector}
\usage{
convert_array(array, to = NULL, ...)
}
\arguments{
\item{array}{A \link[=as_nanoarrow_array]{nanoarrow_array}.}

\item{to}{A target prototype object describing the type to which \code{array}
should be converted, or \code{NULL} to use the default conversion.}

\item{...}{Passed to S3 methods}
}
\value{
An R vector of type \code{to}.
}
\description{
Materialize an Array into an R vector
}
