% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{example_ipc_stream}
\alias{example_ipc_stream}
\title{Example Arrow IPC Data}
\usage{
example_ipc_stream(compression = c("none", "zstd"))
}
\arguments{
\item{compression}{One of "none" or "zstd"}
}
\value{
A raw vector that can be passed to \code{\link[=read_nanoarrow]{read_nanoarrow()}}
}
\description{
An example stream that can be used for testing or examples.
}
\examples{
as.data.frame(read_nanoarrow(example_ipc_stream()))
}
