% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-convert-vector.R
\name{from_nanoarrow_array}
\alias{from_nanoarrow_array}
\alias{infer_nanoarrow_ptype}
\title{Convert an Array to an R vector}
\usage{
from_nanoarrow_array(array, to = NULL, ...)

infer_nanoarrow_ptype(array)
}
\arguments{
\item{array}{A \link[=as_nanoarrow_array]{nanoarrow_array}.}

\item{to}{A target prototype object describing the type to which \code{array}
should be converted, or \code{NULL} to use the default conversion.}

\item{...}{Passed to S3 methods}
}
\value{
An R vector of type \code{to}.
}
\description{
Convert an Array to an R vector
}
