% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-ptype.R
\name{infer_nanoarrow_ptype}
\alias{infer_nanoarrow_ptype}
\title{Infer an R vector prototype}
\usage{
infer_nanoarrow_ptype(x)
}
\arguments{
\item{x}{A \link[=as_nanoarrow_schema]{nanoarrow_schema},
\link[=as_nanoarrow_array]{nanoarrow_array}, or
\link[=as_nanoarrow_array_stream]{nanoarrow_array_stream}.}
}
\value{
An R vector of zero size describing the target into which
the array should be materialized.
}
\description{
Resolves the default \code{ptype} value to use in \code{\link[=materialize_array]{materialize_array()}}.
}
\examples{
infer_nanoarrow_ptype(as_nanoarrow_array(1:10))

}
