% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoarrow-package.R
\docType{package}
\name{nanoarrow-package}
\alias{nanoarrow}
\alias{nanoarrow-package}
\title{nanoarrow: An R Interface to the 'nanoarrow' C Library}
\description{
Provides an R interface to the 'nanoarrow' C library and the 'Apache Arrow' application binary interface. Functions to import and export 'ArrowArray', 'ArrowSchema', and 'ArrowArrayStream' C structures to and from R objects are provided alongside helpers to facilitate zero-copy data transfer among R bindings to libraries implementing the 'Arrow' C data interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apache/arrow-nanoarrow}
  \item Report bugs at \url{https://github.com/apache/arrow-nanoarrow/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Apache Arrow \email{dev@arrow.apache.org} [copyright holder]
}

}
\keyword{internal}
