% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{nanoarrow_buffer_init}
\alias{nanoarrow_buffer_init}
\alias{nanoarrow_buffer_append}
\title{Create and modify nanoarrow buffers}
\usage{
nanoarrow_buffer_init()

nanoarrow_buffer_append(buffer, new_buffer)
}
\arguments{
\item{buffer, new_buffer}{\link[=as_nanoarrow_buffer]{nanoarrow_buffer}s.}
}
\value{
\itemize{
\item \code{nanoarrow_buffer_init()}: An object of class 'nanoarrow_buffer'
\item \code{nanoarrow_buffer_append()}: Returns \code{buffer}, invisibly. Note that
\code{buffer} is modified in place by reference.
}
}
\description{
Create and modify nanoarrow buffers
}
\examples{
buffer <- nanoarrow_buffer_init()
nanoarrow_buffer_append(buffer, 1:5)

array <- nanoarrow_array_modify(
  nanoarrow_array_init(na_int32()),
  list(length = 5, buffers = list(NULL, buffer))
)
as.vector(array)

}
