% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoarrow-package.R
\name{nanoarrow_version}
\alias{nanoarrow_version}
\title{Underlying 'nanoarrow' C library version}
\usage{
nanoarrow_version()
}
\value{
A string identifying the version of nanoarrow this package
was compiled against.
}
\description{
Underlying 'nanoarrow' C library version
}
\examples{
nanoarrow_version()

}
